+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenProjectProperties.java
@@ -21,6 +21,7 @@ import com.google.cloud.tools.jib.event.DefaultEventEmitter;
 import com.google.cloud.tools.jib.event.EventEmitter;
 import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.JibEventType;
+import com.google.cloud.tools.jib.filesystem.AbsoluteUnixPath;
 import com.google.cloud.tools.jib.frontend.JavaLayerConfigurations;
 import com.google.cloud.tools.jib.plugins.common.MainClassInferenceException;
 import com.google.cloud.tools.jib.plugins.common.MainClassResolver;
@@ -56,17 +57,20 @@ public class MavenProjectProperties implements ProjectProperties {
    * @param project the {@link MavenProject} for the plugin.
    * @param log the Maven {@link Log} to log messages during Jib execution
    * @param extraDirectory path to the directory for the extra files layer
+   * @param appRoot root directory in the image where the app will be placed
    * @return a MavenProjectProperties from the given project and logger.
    * @throws MojoExecutionException if no class files are found in the output directory.
    */
-  static MavenProjectProperties getForProject(MavenProject project, Log log, Path extraDirectory)
+  static MavenProjectProperties getForProject(
+      MavenProject project, Log log, Path extraDirectory, AbsoluteUnixPath appRoot)
       throws MojoExecutionException {
     try {
       return new MavenProjectProperties(
           project,
           makeEventEmitter(log),
           new MavenJibLogger(log),
-          MavenLayerConfigurations.getForProject(project, extraDirectory));
+          MavenLayerConfigurations.getForProject(project, extraDirectory, appRoot));
+
     } catch (IOException ex) {
       throw new MojoExecutionException(
           "Obtaining project build output files failed; make sure you have compiled your project "