+++ jib-maven-plugin/src/test/java/com/google/cloud/tools/jib/maven/JibPluginConfigurationTest.java
@@ -59,6 +59,7 @@ public class JibPluginConfigurationTest {
     Assert.assertEquals(
         "<to><auth><password>",
         testPluginConfiguration.getTargetImageAuth().getPasswordPropertyDescriptor());
+    Assert.assertEquals("/app", testPluginConfiguration.getAppRoot());
   }
 
   @Test